make rw and sage support int32

+++ src/cuda/fusion/random_walk.h
@@ -6,7 +6,7 @@ namespace gs {
 namespace impl {
 namespace fusion {
 torch::Tensor FusedRandomWalkCUDA(torch::Tensor seeds, int64_t walk_length,
-                                  int64_t* indices, int64_t* indptr);
+                                  int32_t* indices, int32_t* indptr);
 }
 }  // namespace impl
 }  // namespace gs
